Move TrackItem component into organisms
