+++ src/pages/MostPlayedTracks.tsx
@@ -7,7 +7,7 @@ import API, { Track } from "services/memoir";
 import Footer from "components/Footer";
 import Loading from "components/Loading";
 import Subheader from "components/Subheader";
-import TrackItem from "components/TrackItem";
+import TrackItem from "components/organisms/TrackItem";
 
 interface Props extends RoutableProps {
   api: API;
